requirements.txt
asgiref==3.9.2
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
mysqlclient==2.2.7
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
